+++ backend/src/modules/exams-chat/exams-chat.module.ts
@@ -3,7 +3,11 @@ import { ExamsChatController } from './infrastructure/http/exams.controller';
 import * as AiGenModule from './infrastructure/ai/ai-question.generator';
 import { EXAM_AI_GENERATOR } from './tokens';
 import { GenerateOptionsForQuestionUseCase } from './application/usecases/generate-options-for-question.usecase';
-import { ReinforcementModule } from '../reinforcement/reinforcement.module';
+import { InterviewModule } from '../interviewChat/interview.module';
+import { DsIntService } from '../interviewChat/infrastructure/dsInt.service';
+import { DeepseekAdapter } from '../llm/infrastructure/adapters/ds.adapter';
+import { LLM_PORT } from '../llm/tokens';
+import { PromptTemplateModule } from '../prompt-template/prompt-template.module';
 
 const AiGeneratorClass =
   (AiGenModule as any).AIQuestionGenerator ??
@@ -22,11 +26,14 @@ const AiProvider =
     : { provide: EXAM_AI_GENERATOR, useClass: AiGeneratorFallback };
 
 @Module({
-  imports: [ReinforcementModule],
+  imports: [InterviewModule, PromptTemplateModule],
   controllers: [ExamsChatController],
   providers: [
+    DsIntService,
     AiProvider,
     GenerateOptionsForQuestionUseCase,
+    DeepseekAdapter,
+    { provide: LLM_PORT, useExisting: DeepseekAdapter },
   ],
   exports: [AiProvider],
 })
