--- backend/src/modules/testChat/infrastructure/httpchat/dtoChat/question-response.ts
@@ -1,3 +1,0 @@
-export class QuestionResponse {
-  question: string;
-}
