--- backend/src/modules/testChat/tokens.ts
@@ -1,1 +1,0 @@
-export const ANSWER_AI_GENERATOR = Symbol('ANSWER_AI_GENERATOR');
